create serialize module too
